+++ sdks/python/src/honcho/async_client/client.py
@@ -154,6 +154,9 @@ class AsyncHoncho(BaseModel):
         This method does not make an API call unless `config` or `metadata` is
         provided.
 
+        Provided metadata and configuration will overwrite any existing data in those
+        locations if given.
+
         Args:
             id: Unique identifier for the peer within the workspace. Should be a
             stable identifier that can be used consistently across sessions
@@ -217,6 +220,9 @@ class AsyncHoncho(BaseModel):
         Creates an AsyncSession object that can be used to manage conversations between
         multiple peers. This method does not make an API call unless `config` or
         `metadata` is provided.
+
+        Provided metadata and configuration will overwrite any existing data in those
+        locations if given.
 
         Args:
             id: Unique identifier for the session within the workspace. Should be a